Moved IStringConvertible
