+++ BranchSDK/src/BranchIO/Util/IStringConvertible.h
